requirements.txt
blis==0.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
ChatterBot==1.0.5
click==8.0.1
cymem==2.0.5
fuzzywuzzy==0.18.0
idna==3.2
joblib==1.0.1
mathparse==0.1.2
murmurhash==1.0.5
nltk==3.6.3
numpy==1.21.2
packaging==21.0
Pint==0.17
plac==0.9.6
preshed==2.0.1
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.7.5
python-Levenshtein==0.12.2
pytz==2021.1
PyYAML==5.1.2
regex==2021.8.28
requests==2.26.0
six==1.16.0
spacy==2.1.9
SQLAlchemy==1.2.19
srsly==1.0.5
thinc==7.0.8
tqdm==4.62.3
urllib3==1.26.6
wasabi==0.8.2
